Seperation of UI and data collection, still in progress, but does compile.

+++ ui/SYDE461_UI/CreateNewUser.cs
@@ -26,27 +26,35 @@ namespace SYDE461_UI
 
         private void button1_Click(object sender, EventArgs e)
         {
-            newUser.setUserInfo(textBox1.Text, textBox2.Text);
-            caller.fillUsernameAndPass(this, newUser);
-            
-            //find better way then exposing storedusers like this
-            caller.storedusers.Add(textBox1.Text, textBox2.Text);
-            
-            //add to user file
-            try
-            {
-                //open file. Login screen should have created if not there so append
-                StreamWriter writestream = new StreamWriter("userlist.txt",true);
-                writestream.WriteLine(textBox1.Text);
-                writestream.WriteLine(textBox2.Text);
-                writestream.Close();
-            }
-            catch (Exception ex)
-            {
-                MessageBox.Show(ex.ToString());
-            }
 
-            this.Close();
+            if (caller.storedusers.ContainsKey(textBox1.Text))
+            {
+                MessageBox.Show("Error! User already exists!");
+            }
+            else
+            {
+                newUser.setUserInfo(textBox1.Text, textBox2.Text);
+                caller.fillUsernameAndPass(this, newUser);
+
+                //find better way then exposing storedusers like this
+                caller.storedusers.Add(textBox1.Text, textBox2.Text);
+
+                //add to user file
+                try
+                {
+                    //open file. Login screen should have created if not there so append
+                    StreamWriter writestream = new StreamWriter("userlist.txt", true);
+                    writestream.WriteLine(textBox1.Text);
+                    writestream.WriteLine(textBox2.Text);
+                    writestream.Close();
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show(ex.ToString());
+                }
+
+                this.Close();
+            }
         }
 
         public UserInfo getuserinput()
